Move index_content to Index
Also added dynamic continue to account to log in with Spotify
