+++ package.json
@@ -29,6 +29,7 @@
     "prop-types": "^15.8.1",
     "react": "^18",
     "react-bootstrap": "^2.10.1",
+    "react-cookie": "^7.1.2",
     "react-dom": "^18",
     "react-router-dom": "^6.22.0",
     "react-svg-to-image": "^3.0.0"
